Refactor ComboboxButton in treestructure to new component
