+++ packages/ndla-ui/src/utils/useCombinedRefs.ts
@@ -1,0 +1,30 @@
+/**
+ * Copyright (c) 2022-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import React from 'react';
+import { ForwardedRef, useEffect } from 'react';
+
+const useCombinedRefs = <T extends HTMLButtonElement>(...refs: ForwardedRef<T>[]) => {
+  const targetRef = React.useRef<T | null>(null);
+
+  useEffect(() => {
+    refs.forEach((ref) => {
+      if (!ref) return;
+
+      if (typeof ref === 'function') {
+        ref(targetRef.current);
+      } else {
+        ref.current = targetRef.current;
+      }
+    });
+  }, [refs]);
+
+  return targetRef;
+};
+
+export default useCombinedRefs;
